Fix sorting for books like 1 John (#5)
This change fixes a bug where books whose names contain another book of the bible (e.g. "1 John" containing "John") can get identified as the wrong book and get sorted to the wrong spot in the list.

To fix this, when looking for the book name in a file name, we search from the longest book name to the shortest, to ensure that we don't accidentally associate a file with the wrong book.

I also added some tests to confirm the fix.